+++ assets/asset-discord/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/asset-discord
 
+## 1.0.1-latest.41
+
+### Patch Changes
+
+- sdq
+
 ## 1.0.1-latest.40
 
 ### Patch Changes