+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/shared
 
+## 2.0.1-latest.42
+
+### Patch Changes
+
+- sdq
+- Updated dependencies
+  - @pipelab/migration@1.0.1-latest.41
+
 ## 2.0.1-latest.41
 
 ### Patch Changes