+++ plugins/plugin-electron/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pipelab/plugin-electron
 
+## 1.0.1-latest.36
+
+### Patch Changes
+
+- sdq
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.1-latest.41
+  - @pipelab/constants@1.0.1-latest.41
+  - @pipelab/shared@2.0.1-latest.42
+
 ## 1.0.1-latest.35
 
 ### Patch Changes